Enhanced search query caused by missing ":"
Some movie file name (e.g. Wall Street: Money Never Sleeps) have ":" in their title on yifysubtitles.com, where ":" characters are not allowed for file names in Windows and macOS. This change test different combinations of ":" in the given search query to guess the missing character based on the number of whitespaces in the search query.